discord_bot/requirements.txt
discord.py>=2.0.0
python-dotenv>=0.21.0
aiohttp>=3.8.0
